Refactor dbus API to a reusable client
Creating a new client for every dbus call just wastes resources
